#22967 : Update Empty Starter (#23191)
* #22967 : Update Empty Starter

* #22967 : Adding new Empty Starter with missing portlet.

* Removing test from Unit Test class and moving it to a new Integration Test.

* Delete existing Image and Video Content Types from the original Empty Starter so that the ones in the test bundles are used instead.

* Fixing even more Postman Tests.
