+++ dotCMS/build.gradle
@@ -245,7 +245,7 @@ dependencies {
 
     //// starter config
 
-    starter group: 'com.dotcms', name: 'starter', version: 'empty_20220718', ext: 'zip'
+    starter group: 'com.dotcms', name: 'starter', version: 'empty_20221018', ext: 'zip'
     //Uncomment this line if you want to download the starter that comes with data
     //starter group: 'com.dotcms', name: 'starter', version: '20221010', ext: 'zip'
     testsStarter group: 'com.dotcms', name: 'starter', version: 'empty_20220718', ext: 'zip'
